requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pymisp
asn1crypto==0.24.0
awscli==1.16.20
botocore==1.12.10
certifi==2018.11.29
cffi==1.14.2
chardet==3.0.4
cryptography==2.4.2
idna==2.8
pycparser==2.19
pyOpenSSL==18.0.0
PySocks==1.6.8
requests==2.21.0
requests-futures==0.9.9
rsa==3.4.2
six==1.12.0
urllib3==1.24.2
PyYAML